Organize bot commands into separate modules
